+++ sdk/CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `{TransactionOptions, TransactionOptionsDto}::allow_micro_amount` field;
 - `AddressWithAmount::{new, with_return_address, with_expiration}` methods;
 - `{BaseCoinBalance, BaseCoinBalanceDto}::voting_power` field;
+- `verify_mnemonic()`;
+- `SecretManager::sign_transaction()`;
 
 ### Changed
 
@@ -42,6 +44,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Merge `send_amount` and `send_micro_transaction`;
 - `AddressWithAmount::{address, amount}` fields are no longer public;
 - Fields of `AccountBalance`, `BaseCoinBalance` and `NativeTokensBalance` have been made private and getters have been added;
+- Exposed `AccountParticipationOverview, ParticipationEventWithNodes, AliasOutputOptions, AliasOutputOptionsDto, IncreaseNativeTokenSupplyOptions, IncreaseNativeTokenSupplyOptionsDto, NativeTokenOptions, NativeTokenOptionsDto, NftOptions, NftOptionsDto, OutputOptionsDto` from the `account` module;
+- Made `Wallet::get_bech32_hrp()` public;
 
 ### Removed